PUSH THIS IDK WHY IT WAS NOT WORKING?!
Hey Adam, Idk why but the code was completely missing, as you can see by the video though I did do the assessment, I have no idea the code did not save. If i have to get point deductions because of this that would be unfortunate but I understand. Have a wonderful summer Adam! (I'm assuming Adam is grading this, if not have a wonderful break Veronica or Pito!)